Fill the declaration file with export declaration
Compilation succeeds.

When looking at the trace you see that the resolution is **identic**. So
what happened ?

On the first try, it resolved to the d.ts file but compilation failed
because it was not a valid module : empty file ! When you add the export
declaration it compiles.

Conclusion: declaration files act like "placeholders" for foreign
modules. When importing a module with a **relative**, the compiler will first look for ts
files, if it finds them, it compiles and moves on. If it does not then
it will look for a declaration before looking for a js file (this case
was explored on the previous commits). If this file is valid **and** can
be imported (i.e: exported members correspond to imported ones) then
compilation succeeds, however the js file will not be emitted, unless
you add the `allowJs` compiler option.